ref: rename and move classes to different packages
