+++ src/test/java/fr/insee/pogues/persistence/TestQuestionnaireService.java
@@ -3,8 +3,8 @@ package fr.insee.pogues.persistence;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import fr.insee.pogues.persistence.query.NonUniqueResultException;
-import fr.insee.pogues.persistence.query.QuestionnairesServiceQuery;
+import fr.insee.pogues.persistence.impl.NonUniqueResultException;
+import fr.insee.pogues.persistence.repository.QuestionnaireRepository;
 import fr.insee.pogues.persistence.service.QuestionnairesServiceImpl;
 import fr.insee.pogues.webservice.rest.PoguesException;
 import org.junit.jupiter.api.Test;
@@ -25,7 +25,7 @@ import static org.mockito.Mockito.*;
 class TestQuestionnaireService {
 
     @Mock
-    QuestionnairesServiceQuery questionnairesServiceQuery;
+    QuestionnaireRepository questionnairesServiceQuery;
 
     @InjectMocks
     QuestionnairesServiceImpl questionnairesService;
